fix suggestions
